server/requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
python-dotenv
flask-login
flask-restful
flask-cors
Flask-JWT-Extended